requirements.txt
streamlit
langchain-openai
langchain-google-genai
youtube-transcript-api
pypdf2
python-dotenv
langchain-text-splitters
langchain-google-genai
pytube
langgraph

